Kinematic-model/sample-code/requirements.txt
scipy==1.14.1
numpy==2.1.1
matplotlib==3.9.2
pycairo==1.27.0